fix(security): apply the run-8 audit hardening notes
Four defense-in-depth items from the run-8 audit, none independently
exploitable, on top of the two run-8 fixes (bcdDevice 0x07DD -> 0x07DE).
Each was adversarially re-verified against source before commit.

- credProtect is range-checked. makeCredential accepted a credProtect
  value outside {1,2,3} and stored it verbatim; getAssertion enforces
  protection by exact match, so an out-of-range value silently meant NO
  protection. Now returns CTAP2_ERR_INVALID_OPTION (CTAP 2.1 §12.1).

- hmac-secret-mc empty-salt parity. makeCredential now rejects an
  hmac-secret-mc request with an empty salt up front (MissingParameter),
  matching the getAssertion hmac-secret guard rather than relying only on
  the downstream length check in hmacsecret::eval.

- credentialManagement enumeration counters widened u8 -> u16. skip/total
  and the begin/next counters saturated at 255, so on a fully provisioned
  store (MAX_RESIDENT_CREDENTIALS = 256) the 256th RP/credential was
  invisible to and undeletable via enumeration. Wire encoding is unchanged
  for values <= 255 (canonical CBOR: u16(v) == u8(v)).

- RSA-keygen fast path resets the incoming command chain. The CCID keygen
  fast path already dropped a stale GET RESPONSE tail (clear_pending); it
  now also resets a half-accumulated CLA-0x10 command chain (clear_chaining,
  scrubbing it) so an interrupted chain cannot prepend onto a later command.

Adds +3 regression tests (credProtect range, hmac-secret-mc empty salt,
clear_chaining). Gate green.

+++ crates/rsk-fido/src/credmgmt.rs
@@ -276,8 +276,8 @@ fn enumerate_rps<S: Storage, R: Rng>(
     }
     let target = ctx.state.cm.rp_counter;
 
-    let mut skip = 0u8;
-    let mut total = 0u8;
+    let mut skip = 0u16;
+    let mut total = 0u16;
     let mut found = false;
     let mut rp = [0u8; 256];
     let mut rp_len = 0usize;
@@ -332,7 +332,7 @@ fn enumerate_rps<S: Storage, R: Rng>(
         .map_err(|_| CtapError::Other)?;
     if begin {
         enc.u8(5)
-            .and_then(|e| e.u8(total))
+            .and_then(|e| e.u16(total))
             .map_err(|_| CtapError::Other)?;
     }
     Ok(enc.writer().position())
@@ -354,8 +354,8 @@ fn enumerate_creds<S: Storage, R: Rng>(
     }
     let target = ctx.state.cm.cred_counter;
 
-    let mut skip = 0u8;
-    let mut total = 0u8;
+    let mut skip = 0u16;
+    let mut total = 0u16;
     let mut found = false;
     let mut rec = [0u8; 1024];
     let mut rec_len = 0usize;
@@ -406,7 +406,7 @@ fn enumerate_creds_response(
     rec: &[u8],
     rp_id_hash: &[u8; 32],
     begin: bool,
-    total: u8,
+    total: u16,
     seed: &[u8; 32],
     out: &mut [u8],
 ) -> CtapResult {
@@ -481,7 +481,7 @@ fn enumerate_creds_response(
     // 0x09 totalCredentials — Begin only.
     if begin {
         enc.u8(9)
-            .and_then(|e| e.u8(total))
+            .and_then(|e| e.u16(total))
             .map_err(|_| CtapError::Other)?;
     }
